MOVE FOLDER STRUCTURE
