--- e2e-tests/kitchen-sink/tests/agents.$agentId.spec.ts-snapshots/overall-layout-information-1.aria.yml
@@ -1,11 +1,0 @@
-- navigation:
-    - list:
-        - listitem:
-            - link "Agents":
-                - /url: /agents
-                - img
-        - separator:
-            - img
-        - listitem:
-            - link "Weather Agent":
-                - /url: /agents/weatherAgent
