+++ e2e-tests/kitchen-sink/tests/agents/$agentId/page.spec.ts
