--- e2e-tests/kitchen-sink/tests/workflows.$workflowId.spec.ts-snapshots/overall-layout-information-1.aria.yml
@@ -1,11 +1,0 @@
-- navigation:
-    - list:
-        - listitem:
-            - link "Workflows":
-                - /url: /workflows
-                - img
-        - separator:
-            - img
-        - listitem:
-            - link "complex-workflow":
-                - /url: /workflows/complexWorkflow
